Added first event, activated when the button is pressed
