+++ ContextApp/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
 
     <application android:label="@string/app_name" >
         <activity
-            android:name=".HelloActivity"
+            android:name=".MainActivity"
             android:label="@string/app_name" 
 			android:supportsRtl="true"
 			android:allowBackup="true"
@@ -16,6 +16,13 @@
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity 
+            android:name=".EventButtonPressActivity"
+            android:label="@string/app_name" 
+            android:supportsRtl="true"
+            android:allowBackup="true"
+            android:screenOrientation="landscape">
         </activity>
     </application>
